考试成绩单位移到项目表中

+++ src/com/IotCloud/service/AdminService.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Service;
 
 import com.IotCloud.dao.AdminDao;
 import com.IotCloud.model.Admin;
+import com.IotCloud.util.CommonUtil;
 import com.IotCloud.util.PasswordUtil;
-
 
 @Component("adminService")
 @Service
@@ -18,50 +18,54 @@ public class AdminService {
 	@Autowired
 	private AdminDao adminDao;
 
-	public boolean insertAdmin(String adminId, String userName, String userPasswd, int authority, String orgName, String areaCode) {
-		//不重复添加管理员
+	public boolean insertAdmin(String adminId, String userName, String userPasswd, int authority, String orgName,
+			String areaCode) {
+		// 不重复添加管理员
 		Admin admin = adminDao.getAdminByUserName(userName);
-		if(admin==null) {
+		if (admin == null) {
 			return adminDao.insertAdmin(adminId, userName, userPasswd, authority, orgName, areaCode);
-		}else {
+		} else {
 			return false;
 		}
-		
+
 	}
-	
-//	public boolean deleteAdmin(String userName) {
-//		return adminDao.deleteAdmin(userName);
-//	}
-	
+
+	// public boolean deleteAdmin(String userName) {
+	// return adminDao.deleteAdmin(userName);
+	// }
+
 	public boolean batchDeleteAdmin(List<Admin> adminList) {
+		if (CommonUtil.isEmpty(adminList)) {
+			return true;
+		}
 		return adminDao.deleteAdmins(adminList);
 	}
-	
+
 	public Admin validatePassword(String userName, String inputPasswd) {
 		Admin admin = adminDao.getAdminByUserName(userName);
-		if(admin == null) {
+		if (admin == null) {
 			return null;
 		}
 		String passwdDigest = PasswordUtil.generatePassword(inputPasswd);
-		if(passwdDigest!=null && passwdDigest.equals(admin.getUserPasswd())) {
+		if (passwdDigest != null && passwdDigest.equals(admin.getUserPasswd())) {
 			return admin;
-		}else {
+		} else {
 			return null;
 		}
 	}
-	
+
 	public int updatePassword(String adminId, String oldPasswd, String newPasswd) {
-		if(newPasswd==null || newPasswd.isEmpty()) {
-			//新密码格式不正确
+		if (newPasswd == null || newPasswd.isEmpty()) {
+			// 新密码格式不正确
 			return 13;
 		}
 		return adminDao.updatePassword(adminId, oldPasswd, newPasswd);
 	}
-	
-	public List<Admin> getAdminList(String adminId){
+
+	public List<Admin> getAdminList(String adminId) {
 		return adminDao.getAdminList(adminId);
 	}
-	
+
 	public int resetPassword(String userName) {
 		return adminDao.resetPassword(userName);
 	}
